+++ packages/sdk-release-kit/src/versions/versions-utils.js
@@ -19,7 +19,7 @@ function makePackagesList() {
   const packages = fs.readdirSync(path.join(cwd, '..'))
   return packages.map(pkgPath => {
     const pkgDir = path.join(cwd, '..', pkgPath)
-    const packageJson = require(path.join(pkgDir, 'package.json'))
+    const packageJson = JSON.parse(fs.readFileSync(path.join(pkgDir, 'package.json')))
     return {
       name: packageJson.name,
       path: pkgDir,
@@ -29,7 +29,7 @@ function makePackagesList() {
 
 function verifyDependencies({pkgs, pkgPath, results}) {
   const packageJsonPath = path.resolve(pkgPath, 'package.json')
-  const packageJson = require(packageJsonPath)
+  const packageJson = JSON.parse(fs.readFileSync(packageJsonPath))
   const pkgName = packageJson.name
   const {dependencies} = packageJson
 
@@ -38,7 +38,7 @@ function verifyDependencies({pkgs, pkgPath, results}) {
       const depVersion = dependencies[dep]
       const pkg = pkgs.find(({name}) => name === dep)
       const depPackageJsonPath = path.join(pkg.path, 'package.json')
-      const depPackageJson = require(depPackageJsonPath)
+      const depPackageJson = JSON.parse(fs.readFileSync(depPackageJsonPath))
       const sourceVersion = depPackageJson.version
       results.push({pkgName, dep, depVersion, sourceVersion, error: true})
       verifyDependencies({pkgs, pkgPath: pkg.path, results})
@@ -102,7 +102,7 @@ function findEntryByPackageName(input, target) {
 }
 
 async function checkPackageCommits(pkgPath) {
-  const packageJson = require(path.resolve(pkgPath, 'package.json'))
+  const packageJson = JSON.parse(fs.readFileSync(path.resolve(pkgPath, 'package.json')))
   const {name, version} = packageJson
   const tagName = `${name}@${version}`
   const exclusions = `":(exclude,icase)../*/changelog.md" ":!../*/test/*"`