+++ packages/eyes-protractor/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+
+
+## 1.9.0 - 2020/10/6
+
 - updated to @applitools/eyes-sdk-core@12.2.5 (from 12.2.4)
 - updated to @applitools/visual-grid-client@15.0.5 (from 15.0.4)
 - updated to @applitools/eyes-sdk-core@12.2.6 (from 12.2.5)
 - updated to @applitools/visual-grid-client@15.0.6 (from 15.0.5)
-
+- updated to @applitools/eyes-sdk-core@12.3.0 (from 12.2.6)
+- updated to @applitools/visual-grid-client@15.0.9 (from 15.0.6)
 
 ## 1.8.1 - 2020/9/15